Fix corrupted file
QSW101.3 in quake101.zip is corrupted. when extracting zip file, 7zip says "CRC failed : QSW101.3". If QSW101.3 is used in an attempt to install, the installation will be unable to be completed. This commit replaces QSW101.3 with an identical copy of the one from qsw101_3.zip, which is uncorrupted, and does not have these problems.