Changed the names of the GLQuake versions based on the actual in-game version
The version in the readme doesn't always have the correct version, probably because id didn't bother to update it or made a mistake.

For example, 0.93 has the exact same readme as 0.92 but in-game it's listed as 0.93 in the console.

Also "Glquake vX.XX for Quake vX.XX" isn't correct. It doesn't really make sense, since the DOS Quake update change none of the game files, except for the DOS exe, and GLQauke does not rely on the DOS exe at all.